+++ CHANGELOG.md
@@ -30,6 +30,8 @@
       - libOSRM now creates an own watcher thread then used in shared memory mode to listen for data updates
     - Tools:
       - Added osrm-extract-conditionals tool for checking conditional values in OSM data
+    - Trip Plugin
+      - Added a new feature that finds the optimal route given a list of waypoints, a source and a destination. This does not return a roundtrip and instead returns a one way optimal route from the fixed source to the destination points.
 
 # 5.5.1
   - Changes from 5.5.0